DEVEXP-599 Refactored use of Spark's JacksonParser
Had it duplicated in two places; hoping that the new `JsonRowDeserializer` also sheds light on the important task of converting a JSON string into a Spark row. 

Also addressed another TODO in DocBuilderFactory. 

All refactoring, no functional changes.
