+++ src/main/java/com/marklogic/spark/writer/DocBuilderFactory.java
@@ -44,29 +44,7 @@ class DocBuilderFactory {
     }
 
     DocBuilderFactory withPermissions(String permissionsString) {
-        DocumentMetadataHandle.DocumentPermissions permissions = metadata.getPermissions();
-        // TODO Copied from ml-javaclient-util. Should be in java-client-api.
-        if (permissionsString != null && permissionsString.trim().length() > 0) {
-            String[] tokens = permissionsString.split(",");
-            for (int i = 0; i < tokens.length; i += 2) {
-                String role = tokens[i];
-                if (i + 1 >= tokens.length) {
-                    throw new IllegalArgumentException("Unable to parse permissions string, which must be a comma-separated " +
-                        "list of role names and capabilities - i.e. role1,read,role2,update,role3,execute; string: " + permissionsString);
-                }
-                DocumentMetadataHandle.Capability c;
-                try {
-                    c = DocumentMetadataHandle.Capability.getValueOf(tokens[i + 1]);
-                } catch (Exception e) {
-                    throw new IllegalArgumentException("Unable to parse permissions string: " + permissionsString + "; cause: " + e.getMessage());
-                }
-                if (permissions.containsKey(role)) {
-                    permissions.get(role).add(c);
-                } else {
-                    permissions.add(role, c);
-                }
-            }
-        }
+        metadata.getPermissions().addFromDelimitedString(permissionsString);
         return this;
     }
